Fix EAccessViolation
Fixed EAccessViolation: Access Violation to the list_names array.
This occurs when the program tries to add a new object to the array, but into a redundant index.
Reducing the value of variable 'i' by one can fix it.

Note: Array index usually starts from 0, because the variable 'i' here starts from 1, so I subtract 1 to become 0

Signed-off-by: Aziz Ridhwan <fwindows345@gmail.com>